006/12 11:45

+++ hello/src/main/java/soccer/hello/Repository/PlayerRepository.java
@@ -18,7 +18,7 @@ public interface PlayerRepository {
 
     public List<Player> getAllPlayers();
 
-    public List<Player> findPlayer(Long playerId);
+    public List<Player> findPlayer(Integer playerId);
 
 
 }
