+++ hello/src/main/java/soccer/hello/service/PlayerServi.java
@@ -10,7 +10,7 @@ public interface PlayerServi {
 
     public List<Player> getAllPlayers();
 
-    public List<Player> findPlayer(Long playerId);
+    public List<Player> findPlayer(Integer playerId);
